Add get_instance_pci_request_from_vif
This change extends nova/pci/request.py with a method
that retrieves an instance's PCI request from a given VIF
if the given VIF required a PCI allocation during instance
creation.

The PCI request, if retrieved, belongs to a PCI device
in the compute node where the instance is running.

The change is required to facilitate SR-IOV live migration
allowing to claim VIF related PCI resources on
the destination node.

Change-Id: I9ba475e91b8283f063db446de74d3e4b2de002c5
Partial-Implements: blueprint libvirt-neutron-sriov-livemigration