Support resize/migrate for vGPU instance
First of all, we should claim resources for vGPU(mdev), and then use
the resources prepared before when spawning the instance.

There two phases for claiming resource:
1. At first, claiming vGPU(mdev) resource by new flavor and instance
information on destination host.
2. On the one hand, we should unclaim vGPU resource on destination
host for revert operation. On the other hand, we should unclaim
vGPU resource on source host for confirm operation.

Then we only need to get the previously claimed resources from
memory to spawn the instance on destination(or source) host.

Signed-off-by: Armstrong Liu <liuzz99@chinaunicom.cn>